% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bam_to_csv.R
\name{bam_to_csv}
\alias{bam_to_csv}
\title{Converting BMA to csv for CNV identification.}
\usage{
bam_to_csv(bam_folder, output_folder = NULL)
}
\arguments{
\item{bam_folder}{Path to the folder containing BAM files.}

\item{output_folder}{(optionnal) Path of the output folder.}
}
\value{
A csv file for each BAM, usable by the zscoreProfile function.
}
\description{
This function convert BAM files issued from a Swga sequencing to a specific csv format (chromosome/position/number of reads), rendering it usable by the zscoreProfile function for CNV identification.
}
