% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bam_to_csv_nogff.R
\name{bam_to_csv_nogff}
\alias{bam_to_csv_nogff}
\title{Converting BAM to CSV without GFF, for CNV identification.}
\usage{
bam_to_csv_nogff(bam_folder, output_folder = NULL, min_coverage = 5)
}
\arguments{
\item{bam_folder}{Path to the folder containing BAM files from SWGA.}

\item{output_folder}{(optional) Path to the output folder. Defaults to the working directory if not provided.}

\item{min_coverage}{(optional) Minimum read depth to keep a position. Default is 5.}
}
\value{
A csv file for each BAM, containing only positions with at least \code{min_coverage} reads.
}
\description{
This function converts BAM files from SWGA sequencing into a simple csv format (chromosome / position / number of reads), filtering out low coverage positions. It can be used as input for the cnv_analysis function.
}
