% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cds_cleaning.R
\name{cds_cleaning}
\alias{cds_cleaning}
\title{Removing the sample introns.}
\usage{
cds_cleaning(csv_folder, gff_path, output_folder = NULL)
}
\arguments{
\item{csv_folder}{Path to the folder containing samples CSV files converted from BAM.}

\item{gff_path}{The path to the gff file. It can be downloaded on MalariaGen website.}

\item{output_folder}{(optional) Path to the output folder. Defaults to the working directory if not provided.}
}
\value{
A csv file without introns, usable by the cnv_analysis function.
}
\description{
Remove the introns from the sample in CSV format. It should be used after the \code{bam_to_csv} function for a more accurate analysis.
}
