% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_profile.R
\name{create_profile}
\alias{create_profile}
\title{Profile creation for the zscoreProfile function.}
\usage{
create_profile(profile_csv_folder, output_path = NULL)
}
\arguments{
\item{profile_csv_folder}{A folder containing the sample in csv format (seqnames/pos/count) used to create a new profile.}

\item{output_path}{(optionnal) Name and path of the output file.}
}
\value{
A dataframe representing the SWGA profile of the chosen samples.
}
\description{
This function allow the user to create his own profile to use in the zscoreProfile function.
}
