% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zscoreProfile.R
\name{zscoreProfile}
\alias{zscoreProfile}
\title{CNV identification for SWGA sequencing.}
\usage{
zscoreProfile(csv_folder, region, chr, output_folder = NULL)
}
\arguments{
\item{csv_folder}{A folder containing new sample to compare in csv format.}

\item{region}{The region to compare new sample to. Either a string or a list of strings. ex : c("BENIN", "TOGO")}

\item{chr}{The ID of the chromosome you want to analyse, ex : "Pf3D7_01_v3".}

\item{output_folder}{(optionnal) Name of the output folder.}
}
\value{
A dataframe and a plot containing the score for each gene of each sample.
}
\description{
This function compare a sample to a pre-existing profile in order to identify potential CNV.
}
