% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_profile.R
\name{create_profile}
\alias{create_profile}
\title{Profile creation for the cnv_analysis() function.}
\usage{
create_profile(
  profile_csv_folder,
  chromosomes = 1:14,
  gene_position = NULL,
  output_folder = NULL
)
}
\arguments{
\item{profile_csv_folder}{Path to a folder containing the CSV files. Each file should have 3 columns : \code{seqnames}, \code{pos}, and \code{count}.}

\item{chromosomes}{(optional) Vector of chromosome numbers to analyze (e.g., 1:14 or c(1, 4, 6)). Defaults to 1:14.}

\item{gene_position}{(optional)Path to the file containing each gene to analyze with their start and end. It should have 3 column : \code{gene}, \code{start}, \code{end}. Default to the one provided with the package.}

\item{output_folder}{(optional) Path to the folder where the profile files will be saved. Defaults to the working directory if not provided.}
}
\value{
A dataframe representing the SWGA profile of the chosen samples.
}
\description{
This function allow the user to create his own profile to use in the cnv_analysis function.
}
